Split up loans between views and widgets
