+++ lib/src/features/loans/views/checkout/checkout_page.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:librarian_app/src/features/loans/presentation/checkout/checkout_stepper.dart';
+import 'package:librarian_app/src/features/loans/views/checkout/checkout_stepper.dart';
 
 class CheckoutPage extends StatelessWidget {
   const CheckoutPage({super.key});
