+++ lib/src/features/loans/views/checkout/connected_thing_search_field.dart
