+++ lib/src/features/loans/views/connected_loans_list.dart
