+++ lib/src/features/loans/views/loan_details_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:librarian_app/src/features/loans/data/loans_view_model.dart';
-import 'package:librarian_app/src/features/loans/presentation/loan_details.dart';
+import 'package:librarian_app/src/features/loans/widgets/loan_details.dart';
 import 'package:provider/provider.dart';
 
 import '../../dashboard/views/mobile_layout.dart';
