+++ lib/src/features/loans/views/searchable_loans_list.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:librarian_app/src/features/common/widgets/submit_text_field.dart';
 import 'package:librarian_app/src/features/loans/data/loan_model.dart';
-import 'package:librarian_app/src/features/loans/presentation/connected_loans_list.dart';
+import 'package:librarian_app/src/features/loans/views/connected_loans_list.dart';
 
 class SearchableLoansList extends StatefulWidget {
   final Function(LoanModel)? onLoanTapped;
