+++ lib/src/features/loans/views/things_list_view.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:librarian_app/src/features/loans/data/things_view_model.dart';
-import 'package:librarian_app/src/features/loans/presentation/thing_list_tile.dart';
+import 'package:librarian_app/src/features/loans/widgets/thing_list_tile.dart';
 import 'package:provider/provider.dart';
 
 import '../data/thing_model.dart';
