+++ lib/src/features/loans/widgets/loan_details.dart
