+++ lib/src/features/loans/widgets/thing_list_tile.dart
